+++ tests/Lithuania/FathersDayTest.php
@@ -1,0 +1,83 @@
+<?php
+
+declare(strict_types = 1);
+
+/**
+ * This file is part of the 'Yasumi' package.
+ *
+ * The easy PHP Library for calculating holidays.
+ *
+ * Copyright (c) 2015 - 2025 AzuyaLabs
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Sacha Telgenhof <me at sachatelgenhof dot com>
+ */
+
+namespace Yasumi\tests\Lithuania;
+
+use Yasumi\Holiday;
+use Yasumi\Provider\Lithuania;
+use Yasumi\tests\HolidayTestCase;
+
+/**
+ * Class containing tests for Father's Day in Lithuania.
+ *
+ * @author Gedas Lukošius <gedas@lukosius.me>
+ */
+class FathersDayTest extends LithuaniaBaseTestCase implements HolidayTestCase
+{
+    /**
+     * The name of the holiday to be tested.
+     */
+    public const HOLIDAY = 'fathersDay';
+
+    /**
+     * Test if holiday is not defined before restoration.
+     *
+     * @throws \Exception
+     */
+    public function testHoliday(): void
+    {
+        $year = $this->generateRandomYear();
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            $year,
+            new \DateTime("first sunday of june {$year}", new \DateTimeZone(self::TIMEZONE))
+        );
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testTranslation(): void
+    {
+        $this->assertTranslatedHolidayName(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(),
+            [self::LOCALE => 'Tėvo dieną']
+        );
+        $this->assertTranslatedHolidayName(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(),
+            ['en' => 'Father’s Day']
+        );
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testHolidayType(): void
+    {
+        $this->assertHolidayType(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(),
+            Holiday::TYPE_OFFICIAL
+        );
+    }
+}